Device mock as separate utility
